use init from @angular/cli instead of override it
