+++ bin/cliberry
@@ -22,4 +22,5 @@ if (Number(version[0]) < 8 || (Number(version[0]) === 8 && Number(version[1]) < 
   process.exit(3);
 }
 
-require('../lib/init');
+require('../lib/cli-parameter');
+require('@angular/cli/lib/init');
